refactor: Moved graph capturing sheet
